Added logout, added memory of a user's packages, made page system better
Pressing the logout button logs out a user to the logout page and clears
all cookies and then redirects them to the login page. Going to the home
page redirects to the login page. If a user adds a uuid, that package
will always stay on the trackPackages screen until that user logs out.
This includes if the user inputs multiple uuids.